🚚 Move coins logic into own module
